+++ custom_components/topomation/const.py
@@ -2,7 +2,7 @@
 
 DOMAIN = "topomation"
 NAME = "Topomation"
-VERSION = "0.1.18"
+VERSION = "0.1.19"
 
 # Storage
 STORAGE_VERSION = 1